% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caoChIP}
\alias{caoChIP}
\title{caoChIP: serialized GenomicFiles instance}
\format{GenomicFiles instance}
\source{
NCBI GEO GSE65516
}
\usage{
caoChIP
}
\description{
caoChIP: serialized GenomicFiles instance
}
\examples{
data(caoChIP)
caoChIP
}
\keyword{datasets}
