% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shin.R
\name{mowchiBrowse}
\alias{mowchiBrowse}
\title{visualize coverage profiles for selected genes and MOWChIP protocols}
\usage{
mowchiBrowse()
}
\description{
visualize coverage profiles for selected genes and MOWChIP protocols
}
\note{
This function sets up references to BigWig files in a public
AWS S3 bucket and will use user selections to extract relevant coverage
scores.
}
