% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseMOWChIP.R
\name{mowChInS3}
\alias{mowChInS3}
\title{create HTTP paths to S3 bucket holding selected MOWChIP bigwig files}
\usage{
mowChInS3()
}
\description{
create HTTP paths to S3 bucket holding selected MOWChIP bigwig files
}
\examples{
mowChInS3()
}
