% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseMOWChIP.R
\name{viewByGene}
\alias{viewByGene}
\title{multimodal Gviz-based visualization of MOWChIP tracks}
\usage{
viewByGene(gf = caoChIP[, 1:6], sym = "IGHA2", radius = 1e+05, gstr,
  namegen = function(x) x$numCells[1], ...)
}
\arguments{
\item{gf}{GenomicFiles instance giving paths to .bw files}

\item{sym}{gene symbol}

\item{radius}{number of bp around coding region to display}

\item{gstr}{a GRanges with gene annotation information, e.g., genes(EnsDb.Hsapiens.v75)}

\item{namegen}{a function of one argument assumed to be a row of \code{colData(gf)} that will be used as the \code{name} parameter for the
\code{\link[Gviz]{DataTrack}} derived from each sample}

\item{\dots}{passed to \code{\link[Gviz]{plotTracks}} exclusive of type which is set to "polygon"}
}
\description{
multimodal Gviz-based visualization of MOWChIP tracks
}
\examples{
data(caoChIP)
S3paths = mowChInS3()
if (interactive()) {
 remoteMowK4me3 = GenomicFiles(files=S3paths[1:4], 
    colData=colData(caoChIP[,c(1,3,5,7)]))
 remoteMowK27ac = GenomicFiles(files=S3paths[5:8], 
    colData=colData(caoChIP[,c(9,11,13,15)]))
 require(EnsDb.Hsapiens.v75)
 gg = genes(EnsDb.Hsapiens.v75)
 require(GenomeInfoDb)
 seqlevelsStyle(gg) = "UCSC"
 viewByGene(gf=remoteMowK4me3, sym="IGHA2", radius=80000, gstr=gg)
 viewByGene(gf=remoteMowK27ac, sym="SPI1", radius=80000, gstr=gg)
 }
}
